Use int type for saddr size

+++ src/spade/reporter/audit/kernel-modules/spade/audit/kernel/helper/sock.h
@@ -28,7 +28,7 @@ struct kernel_helper_sock_saddr_info
 {
     int fd;
     struct sockaddr_storage saddr;
-    size_t saddr_size;
+    int saddr_size;
     short sock_type;
     unsigned int net_ns_inum;
 };
@@ -77,7 +77,7 @@ int kernel_helper_sock_get_saddr_info_from_fd(
         -ive    -> Error code.
 */
 int kernel_helper_sock_copy_sock_len_from_userspace(
-    uint32_t *dst_len, uint32_t __user *src_len
+    uint32_t *dst_len, int __user *src_len
 );
 
 /*
@@ -94,8 +94,8 @@ int kernel_helper_sock_copy_sock_len_from_userspace(
         -ive    -> Error code.
 */
 int kernel_helper_sock_copy_saddr_from_userspace(
-    struct sockaddr_storage *dst, uint32_t avail_dst_len,
-    const struct sockaddr __user *src, uint32_t src_len
+    struct sockaddr_storage *dst, int avail_dst_len,
+    const struct sockaddr __user *src, int src_len
 );
 
 /*
@@ -111,7 +111,7 @@ int kernel_helper_sock_copy_saddr_from_userspace(
         -ive    -> Error code.
 */
 int kernel_helper_sock_copy_saddr_in_msghdr_from_userspace(
-    struct sockaddr_storage *dst, uint32_t *dst_len,
+    struct sockaddr_storage *dst, int *dst_len,
     struct msghdr __user *src
 );
